Return to working version
