+++ src/store/models/network.ts
@@ -1,104 +1,40 @@
 import { createJSONStorage, devtools, persist } from 'zustand/middleware';
-import { immer } from 'zustand/middleware/immer';
 import { createStore } from 'zustand/vanilla';
-import { CurrentNetworkType, NetworkType } from '../../types/network.types';
-import { GetSetType } from './models.types';
+import { EnvironmentsEnum } from '../../types/enums.types';
 
-export const defaultNetwork: CurrentNetworkType = {
-  id: 'not-configured',
-  chainId: '',
-  name: 'NOT CONFIGURED',
-  egldLabel: '',
-  decimals: '18',
-  digits: '4',
-  gasPerDataByte: '1500',
-  walletConnectDeepLink: '',
-  walletConnectBridgeAddress: '',
-  walletConnectV2RelayAddress: '',
-  walletConnectV2ProjectId: '',
-  walletConnectV2Options: {},
-  walletAddress: '',
-  apiAddress: '',
-  explorerAddress: '',
-  apiTimeout: '4000'
-};
-
-function getRandomAddressFromNetwork(walletConnectAddresses: string[]) {
-  return walletConnectAddresses[
-    Math.floor(Math.random() * walletConnectAddresses.length)
-  ];
-}
-
-export const namespace = 'network';
-
-export enum KeysEnum {
-  network = 'network',
+export enum NetworkKeysEnum {
   chainID = 'chainID',
-  customWalletAddress = 'customWalletAddress',
-  setChainID = 'setChainID',
-  setCustomWalletAddress = 'setCustomWalletAddress',
-  initializeNetworkConfig = 'initializeNetworkConfig'
+  setChainID = 'setChainID'
 }
 
-export const initialState = {
-  [namespace]: {
-    network: defaultNetwork,
-    chainID: '-1',
-    setChainID: (_chainID: string) => {},
-    setCustomWalletAddress: (_customWalletAddress: string) => {},
-    initializeNetworkConfig: (_network: NetworkType) => {}
-  }
+export type NetworkRootState = {
+  [NetworkKeysEnum.chainID]: EnvironmentsEnum;
+  [NetworkKeysEnum.setChainID]: (env: EnvironmentsEnum) => void;
 };
 
-export type RootState = typeof initialState;
+export const networkStoreDefinition = (
+  set: (
+    partial:
+      | NetworkRootState
+      | Partial<NetworkRootState>
+      | ((
+          state: NetworkRootState
+        ) => NetworkRootState | Partial<NetworkRootState>),
+    replace?: boolean | undefined
+  ) => void
+): NetworkRootState => ({
+  [NetworkKeysEnum.chainID]: EnvironmentsEnum.devnet,
+  [NetworkKeysEnum.setChainID]: (env: EnvironmentsEnum) =>
+    set((state) => ({
+      ...state,
+      chainID: env
+    }))
+});
 
-export const definition = (set: GetSetType<RootState>) => {
-  const values: RootState['network'] = {
-    network: defaultNetwork,
-    chainID: '-1',
-    initializeNetworkConfig: (newNetwork) => {
-      const walletConnectV2RelayAddress = getRandomAddressFromNetwork(
-        newNetwork.walletConnectV2RelayAddresses
-      );
-      const { walletConnectV2RelayAddresses, ...rest } = newNetwork;
-      return set(
-        ({ network }) => {
-          network[KeysEnum.network] = {
-            ...network[KeysEnum.network],
-            ...rest,
-            walletConnectV2RelayAddress
-          };
-        },
-        false,
-        { type: KeysEnum.initializeNetworkConfig }
-      );
-    },
-    setCustomWalletAddress: (customWalletAddress) =>
-      set(
-        ({ network }) => {
-          network[KeysEnum.network].customWalletAddress = customWalletAddress;
-        },
-        false,
-        { type: KeysEnum.setCustomWalletAddress }
-      ),
-    setChainID: (chainID) =>
-      set(
-        ({ network }) => {
-          network[KeysEnum.chainID] = chainID;
-        },
-        false,
-        { type: KeysEnum.setChainID }
-      )
-  };
-  return {
-    [namespace]: values
-  };
-};
-
-export const store = createStore<RootState>()(
+export const sessionNetworkStore = createStore<NetworkRootState>()(
   devtools(
-    persist(immer(definition), {
-      name: 'networkStore',
+    persist(networkStoreDefinition, {
+      name: 'sessionNetworkStore',
       storage: createJSONStorage(() => sessionStorage)
     })
   )
